BINIUS-526: Move build_g and build_monster_segments onto their owning types
Both functions already took one type as their main argument, so this
turns them into methods on that type instead of free functions the
caller has to locate in a sibling module. Bodies, generics, and
algorithms are unchanged; only the receiver moves from an explicit
argument to `self`. Updates all three call sites, across the
single-instance prover, the batched M4 prover, and the shift-reduction
benchmark.
