+++ crates/prover/src/protocols/shift/monster.rs
@@ -5,24 +5,11 @@ use std::iter;
 
 use binius_compute::{Allocator, VecLike};
 use binius_core::{ShiftVariant, constraint_system::Shift, word::Word};
-use binius_field::{BinaryField, Field, PackedField, WideMul};
+use binius_field::{BinaryField, Field, PackedField};
 use binius_math::{FieldBuffer, FieldVec, multilinear::eq::eq_ind_partial_eval};
-use binius_utils::{
-	checked_arithmetics::log2_ceil_usize,
-	rayon::{
-		prelude::*,
-		task_size::{IndexedParallelIteratorExt, WorkPerItem},
-	},
-};
-use binius_verifier::protocols::shift::{BINMUL_ARITY, BITAND_ARITY, INTMUL_ARITY, ZERO_ARITY};
 use tracing::instrument;
 
-use super::{
-	claims::PreparedOperatorClaims,
-	key_collection::{DenseShiftEncoding, KeyCollection, KeySegment, Operation},
-	phase_1::SHIFT_OPERATOR_LOG_LEN,
-	shift_ind::ShiftChallenge,
-};
+use super::{phase_1::SHIFT_OPERATOR_LOG_LEN, shift_ind::ShiftChallenge};
 
 /// The width the half-word (`*32`) shift variants act over.
 const HALF_WORD_BITS: usize = 32;
@@ -31,11 +18,11 @@ const HALF_WORD_BITS: usize = 32;
 ///
 /// A table holds the `arity` weights of each shift sequence the segment uses, in dense shift index
 /// order, so a key's weights are the chunk at `key.dense_shift_idx * arity`.
-struct ScalarTables<F> {
-	zero: Vec<F>,
-	bitand: Vec<F>,
-	intmul: Vec<F>,
-	binmul: Vec<F>,
+pub(super) struct ScalarTables<F> {
+	pub(super) zero: Vec<F>,
+	pub(super) bitand: Vec<F>,
+	pub(super) intmul: Vec<F>,
+	pub(super) binmul: Vec<F>,
 }
 
 /// The equality-indicator weights of a shift sequence's outer slot.
@@ -43,7 +30,7 @@ struct ScalarTables<F> {
 /// The sequence weight factorizes across its two slots, so each slot carries its own two tensors:
 /// one over the variant axis, one over the amount axis.
 /// Keeping them apart holds the weights at `2 * SHIFT_COUNT` entries rather than `SHIFT_COUNT^2`.
-struct OuterSlotWeights<F: Field> {
+pub(super) struct OuterSlotWeights<F: Field> {
 	/// The equality indicator over the outer variant axis, one weight per shift variant.
 	variant: FieldBuffer<F>,
 	/// The equality indicator over the outer amount axis, one weight per shift amount.
@@ -52,7 +39,7 @@ struct OuterSlotWeights<F: Field> {
 
 impl<F: Field> OuterSlotWeights<F> {
 	/// The equality indicators of the outer slot's challenge point.
-	fn new(outer: &ShiftChallenge<F>) -> Self {
+	pub(super) fn new(outer: &ShiftChallenge<F>) -> Self {
 		Self {
 			variant: eq_ind_partial_eval::<F>(&outer.variant),
 			amount: eq_ind_partial_eval::<F>(&outer.amount),
@@ -61,7 +48,7 @@ impl<F: Field> OuterSlotWeights<F> {
 
 	/// The weight one outer shift contributes to its sequence's scalar.
 	#[inline]
-	fn weight(&self, shift: Shift) -> F {
+	pub(super) fn weight(&self, shift: Shift) -> F {
 		self.variant.as_ref()[shift.variant as usize] * self.amount.as_ref()[shift.amount as usize]
 	}
 }
@@ -223,163 +210,6 @@ where
 	unsafe { values.set_len(packed_len) };
 
 	FieldBuffer::new(SHIFT_OPERATOR_LOG_LEN, values)
-}
-
-/// Constructs the "monster multilinear" that combines all shift operations into a single
-/// multilinear.
-///
-/// This function builds a comprehensive multilinear polynomial that encapsulates the AND, IMUL and
-/// BMUL constraints with their associated shift operations. For each witness word, it computes the
-/// contribution from all constraints involving that word, weighted by the h evaluation and lambda
-/// powers.
-///
-/// # Construction Process
-///
-/// 1. **Compute lambda powers**: Powers λ^(i+1) for each operand index in both operations
-/// 2. **Build scalar matrix**: Create scalars combining lambda powers, the h evaluation, and the
-///    `r_s` and `r_v` tensors
-/// 3. **Process keys in parallel**: For each word, accumulate contributions from all its
-///    constraints
-///
-/// # Formula
-///
-/// For each word w, computes:
-/// ```text
-/// ∑_{key ∈ keys[w]} key.accumulate(constraint_indices, tensor, scalars[key.dense_shift_idx])
-/// ```
-/// where `scalars[key.dense_shift_idx]` is the contiguous per-operand chunk encoding
-/// `λ^(operand_idx+1) × h_eval × r_v_tensor[shift_variant] × r_s_tensor[shift_amount]` for operand
-/// index `operand_idx`, and `(shift_variant, shift_amount)` the pair the key's segment decodes its
-/// dense shift index to. The h evaluation comes from
-/// [`ShiftIndSumcheck`](super::ShiftIndSumcheck), which proves it.
-///
-/// # Usage
-///
-/// Used in phase 2 of the shift protocol. The two returned buffers are the segments of the
-/// witness's monster multilinear: the public piece over `log_public_words` variables and the
-/// hidden piece over `log_witness_words` variables (the hidden words at the base, zeros above).
-/// The sparse first sumcheck round consumes them without materializing the combined buffer.
-#[instrument(skip_all, name = "build_monster_segments")]
-#[allow(clippy::too_many_arguments)]
-pub fn build_monster_segments<F, P: PackedField<Scalar = F>, A: Allocator>(
-	alloc: &A,
-	key_collection: &KeyCollection,
-	prepared: &PreparedOperatorClaims<F>,
-	h_eval: F,
-	inner: &ShiftChallenge<F>,
-	outer: &ShiftChallenge<F>,
-) -> (FieldVec<P, A>, FieldVec<P, A>)
-where
-	F: BinaryField,
-{
-	let r_v_tensor = eq_ind_partial_eval::<F>(&inner.variant);
-	let r_s_tensor = eq_ind_partial_eval::<F>(&inner.amount);
-
-	// Invariant: a key's sequence weight factorizes across its two slots.
-	//
-	//     eq(r_v1, v_1) * eq(r_s1, s_1)  *  eq(r_v2, v_2) * eq(r_s2, s_2)
-	//     \______ inner slot _________/     \______ outer slot ________/
-	//
-	// So one table per slot suffices, at `2 * SHIFT_COUNT` entries instead of `SHIFT_COUNT^2`.
-	let outer_weights = OuterSlotWeights::<F>::new(outer);
-
-	// The scalars of one operation, laid out with the operand index innermost so that the `arity`
-	// weights for one `key.dense_shift_idx` form a contiguous chunk that [`Key::accumulate_wide`]
-	// can index directly by operand index.
-	//
-	// A key's sequence selects itself through an equality indicator over both slots' axes.
-	// The h evaluation is one factor shared by every key of the operation.
-	let build_scalars = |arity: usize,
-	                     lambda_powers: &[F],
-	                     dense_shift_enc: &DenseShiftEncoding| {
-		let mut scalars = vec![F::ZERO; arity * dense_shift_enc.len()];
-		for (dense_shift_idx, [inner_shift, outer_shift]) in dense_shift_enc.iter().enumerate() {
-			let shift_scalar = h_eval
-				* r_v_tensor.as_ref()[inner_shift.variant as usize]
-				* r_s_tensor.as_ref()[inner_shift.amount as usize]
-				* outer_weights.weight(outer_shift);
-			for operand_idx in 0..arity {
-				scalars[dense_shift_idx * arity + operand_idx] =
-					lambda_powers[operand_idx] * shift_scalar;
-			}
-		}
-		scalars
-	};
-
-	// Each segment has its own dense shift encoding, so it has its own scalar tables.
-	let build_scalar_tables = |dense_shift_enc: &DenseShiftEncoding| ScalarTables {
-		zero: build_scalars(ZERO_ARITY, &prepared.zero.lambda_powers, dense_shift_enc),
-		bitand: build_scalars(BITAND_ARITY, &prepared.bitand.lambda_powers, dense_shift_enc),
-		intmul: build_scalars(INTMUL_ARITY, &prepared.intmul.lambda_powers, dense_shift_enc),
-		binmul: build_scalars(BINMUL_ARITY, &prepared.binmul.lambda_powers, dense_shift_enc),
-	};
-
-	// The scalar for one word of a segment: the accumulated contribution of all its keys. The
-	// per-key wide accumulations are summed unreduced and reduced once at the end.
-	let word_scalar = |segment: &KeySegment, tables: &ScalarTables<F>, index: usize| {
-		let wide = segment
-			.word_keys(index)
-			.iter()
-			.map(|key| {
-				// The scalar table is per operation, and its stride is that operation's arity.
-				let (scalars, arity) = match key.operation {
-					Operation::Zero => (&tables.zero, ZERO_ARITY),
-					Operation::BitwiseAnd => (&tables.bitand, BITAND_ARITY),
-					Operation::IntegerMul => (&tables.intmul, INTMUL_ARITY),
-					Operation::BinMul => (&tables.binmul, BINMUL_ARITY),
-				};
-				let base = key.dense_shift_idx as usize * arity;
-				key.accumulate_wide(
-					&segment.constraint_indices,
-					prepared[key.operation].r_x_prime_tensor.as_ref(),
-					&scalars[base..base + arity],
-				)
-			})
-			.sum::<<F as WideMul>::Output>();
-		F::reduce(wide)
-	};
-
-	// Each segment sits at the base of its buffer: the public piece fills its power-of-two
-	// length exactly, the hidden piece is zero-padded up to the hidden segment length.
-	let build_segment = |segment: &KeySegment, log_len: usize| {
-		let tables = build_scalar_tables(&segment.dense_shift_enc);
-		let capacity = 1 << log_len.saturating_sub(P::LOG_WIDTH);
-		let n_words = segment.n_words();
-		// Full packed elements: each maps exactly `P::WIDTH` words, so `from_scalars` sees a
-		// statically-sized iterator. The trailing partial element is filled separately below.
-		let n_full = n_words / P::WIDTH;
-		// Allocate the backing buffer up front from the allocator, then fill the `n_full` aligned
-		// packed elements in parallel through its spare capacity — the single allocation happens
-		// before the parallel region, which only writes.
-		let mut values = alloc.alloc::<P>(capacity);
-		values.spare_capacity_mut()[..n_full]
-			.par_iter_mut()
-			.enumerate()
-			.with_min_task(WorkPerItem::FieldMuls)
-			.for_each(|(chunk_index, slot)| {
-				let start = chunk_index * P::WIDTH;
-				slot.write(P::from_scalars(
-					(0..P::WIDTH).map(|i| word_scalar(segment, &tables, start + i)),
-				));
-			});
-		// Safety: the parallel loop above initialized every one of the `n_full` slots.
-		unsafe { values.set_len(n_full) };
-		if !n_words.is_multiple_of(P::WIDTH) {
-			let start = n_full * P::WIDTH;
-			values.push(P::from_scalars(
-				(start..n_words).map(|word_index| word_scalar(segment, &tables, word_index)),
-			));
-		}
-		values.resize(capacity, P::default());
-		FieldBuffer::new(log_len, values)
-	};
-
-	// The segment word count need not be a power of two; the monster spans the rounded-up count.
-	let log_public_words = log2_ceil_usize(key_collection.public.n_words());
-	let public_monster = build_segment(&key_collection.public, log_public_words);
-	let hidden_monster = build_segment(&key_collection.hidden, key_collection.log_witness_words());
-
-	(public_monster, hidden_monster)
 }
 
 #[cfg(test)]
