+++ crates/prover/src/protocols/shift/phase_1.rs
@@ -1,10 +1,7 @@
 // Copyright 2025 Irreducible Inc.
 // Copyright 2026 The Binius Developers
 
-use std::{
-	iter,
-	ops::{Deref, Range},
-};
+use std::{iter, ops::Deref};
 
 use binius_compute::Allocator;
 use binius_core::word::Word;
@@ -19,20 +16,11 @@ use binius_ip_prover::{
 	},
 };
 use binius_math::{FieldBuffer, FieldVec, multilinear::fold::fold_highest_var_inplace};
-use binius_utils::rayon::{
-	prelude::*,
-	task_size::{IndexedParallelIteratorExt, WorkPerItem},
-};
 use binius_verifier::protocols::shift::LOG_SHIFT_COUNT;
-use bytemuck::zeroed_vec;
-use itertools::izip;
 use tracing::instrument;
 
 use super::{
-	claims::PreparedOperatorClaims,
-	key_collection::{DenseShiftEncoding, KeySegment},
-	monster::shift_operator_table,
-	outer::OuterShiftStage,
+	key_collection::DenseShiftEncoding, monster::shift_operator_table, outer::OuterShiftStage,
 	shift_ind::ShiftChallenge,
 };
 
@@ -79,7 +67,7 @@ pub struct Phase1Output<F> {
 }
 
 /// The number of packed elements one row of `Word::BITS` scalars occupies.
-const fn row_len<P: PackedField>() -> usize {
+pub(super) const fn row_len<P: PackedField>() -> usize {
 	assert!(
 		P::LOG_WIDTH <= Word::LOG_BITS,
 		"a row of `Word::BITS` scalars must be a whole number of packed elements"
@@ -547,133 +535,6 @@ impl<A: Allocator, F: Field, P: PackedField<Scalar = F>> SumcheckProver<F>
 			Stage::Bits(prover) => prover.finish(),
 		}
 	}
-}
-
-/// Accumulates one key segment's rows of the witness-and-batching multilinear.
-///
-/// Every witness word can carry several shift keys, one per operand position it feeds.
-/// For each key, this folds its constraint-index tensor and batching powers into one
-/// accumulator value, then scatters that value across the bit positions the word has set.
-///
-/// # Returns
-///
-/// One row of weights per shift the segment uses, one weight per bit position of a word, at
-/// the position the segment's own dense encoding assigns to that shift.
-#[instrument(skip_all, name = "build_g")]
-pub fn build_g<F: Field, P: PackedField<Scalar = F>>(
-	words: &[Word],
-	segment: &KeySegment,
-	prepared: &PreparedOperatorClaims<F>,
-) -> Box<[P]> {
-	// One row of `Word::BITS` scalars per shift the segment uses.
-	let row_len = row_len::<P>();
-	let acc_size = segment.dense_shift_enc.len() * row_len;
-
-	const {
-		assert!(
-			P::WIDTH <= 8,
-			"the optimizations below work only when the width of `P` is less than 8 (which is true for all packed 128b fields we use for now)"
-		);
-	}
-
-	// Precompute once: a map from an 8-bit mask (one bit per candidate lane) to the packed
-	// lane-selection mask it names.
-	// Every accumulator below reuses this table instead of rebuilding a mask per word.
-	let packed_masks_map = (0..1 << P::WIDTH)
-		.map(|i| P::make_mask((0..P::WIDTH).map(|bit_index| (i >> bit_index) & 1 == 1)))
-		.collect::<Vec<_>>();
-	// A mask for the low bits that actually address a lane.
-	let low_bits_mask = (1u8 << P::WIDTH) - 1;
-
-	// Each word carries the keys its segment-relative range names.
-	//
-	// The fold runs in parallel.
-	// Each task owns a chunk of words and its own accumulator.
-	// The accumulators merge at the end.
-	words
-		.par_iter()
-		.zip(segment.key_ranges.par_iter())
-		.with_min_task(WorkPerItem::FieldMuls)
-		.fold(
-			|| zeroed_vec::<P>(acc_size).into_boxed_slice(),
-			|mut multilinears, (word, Range { start, end })| {
-				let keys = &segment.keys[*start as usize..*end as usize];
-
-				// A word can carry several keys, one per operand position it feeds.
-				for key in keys {
-					let operator_data = &prepared[key.operation];
-
-					// Fold this key's accumulator value: the constraint-index tensor
-					// weighted by the batching powers for this operand position.
-					let acc = key.accumulate(
-						&segment.constraint_indices,
-						operator_data.r_x_prime_tensor.as_ref(),
-						&operator_data.lambda_powers,
-					);
-					let acc_packed = P::broadcast(acc);
-
-					// Scatter the accumulator value into every bit position where the word
-					// is set to 1.
-					//
-					// A naive version would loop over all `Word::BITS` positions and test
-					// each bit in turn.
-					//
-					// This instead walks only the word's nonzero bytes.
-					// Inside each byte, it selects every set bit's lane in one packed
-					// operation, using the precomputed mask above.
-					debug_assert!(
-						(key.dense_shift_idx as usize) < segment.dense_shift_enc.len(),
-						"a key indexes the dense shift encoding of its own segment"
-					);
-					let start = key.dense_shift_idx as usize * row_len;
-					let values = &mut multilinears[start..start + row_len];
-					let values_per_byte = Word::BYTES >> P::LOG_WIDTH;
-					let mut remaining_word = word.0;
-					let mut byte_index = 0;
-					while remaining_word != 0 {
-						let byte = remaining_word as u8;
-						let byte_values =
-							&mut values[byte_index * values_per_byte..][..values_per_byte];
-						for value_index in 0..(8 >> P::LOG_WIDTH) {
-							unsafe {
-								let packed_mask_index =
-									((byte >> (value_index * P::WIDTH)) & low_bits_mask) as usize;
-
-								// Safety:
-								// - `packed_masks_map` holds one entry per possible `P::WIDTH`-bit
-								//   value, so this index always fits.
-								let packed_mask = packed_masks_map.get_unchecked(packed_mask_index);
-
-								// Safety:
-								// - `values` spans `(8 >> P::LOG_WIDTH)` elements after the
-								//   chunking above.
-								// - `value_index` stays within that range by construction.
-								*byte_values.get_unchecked_mut(value_index) +=
-									acc_packed.select(packed_mask);
-							}
-						}
-						remaining_word >>= 8;
-						byte_index += 1;
-					}
-				}
-
-				multilinears
-			},
-		)
-		// Merge every task's accumulator pairwise.
-		//
-		// A merge seeded with a partial that already exists never touches a buffer of zeros.
-		// An identity element instead would allocate and zero one accumulator per merge.
-		// Then it would add all of it for nothing.
-		.reduce_with(|mut acc, local| {
-			izip!(acc.iter_mut(), local.iter()).for_each(|(acc, local)| {
-				*acc += *local;
-			});
-			acc
-		})
-		// An empty word list produces no partial accumulator at all.
-		// So its rows are zero.
-		.unwrap_or_else(|| zeroed_vec::<P>(acc_size).into_boxed_slice())
 }
 
 #[cfg(test)]
